% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acid_Base.R
\name{balance_ions}
\alias{balance_ions}
\title{Ion balance}
\usage{
balance_ions(
  Temp0 = 20,
  Alk0,
  pH0,
  TDS = NA,
  Na.Conc,
  Ca.Conc,
  Mg.Conc,
  K.Conc = 0,
  Cl.Conc,
  SO4.Conc,
  ...
)
}
\arguments{
\item{Temp0}{temperature in degrees C (optional, defaults to 20)}

\item{Alk0}{alkalinity in mg CaCO3/L}

\item{pH0}{pH in SU}

\item{TDS}{total dissolved solids (TDS) measured (optional, but necessary if you want the difference between calculated and measured)}

\item{Na.Conc}{sodium concentration in mg/L}

\item{Ca.Conc}{calcium concentration mg/L}

\item{Mg.Conc}{magnesium concentration mg/L}

\item{K.Conc}{potassium concentration mg/L (optional, defaults to 0)}

\item{Cl.Conc}{chloride concentration mg/L}

\item{SO4.Conc}{sulfate concentration mg/L}
}
\description{
This function takes ions, converts them to mg as CaCO3/L and determines the balance.
It can also calculate the difference between the measured TDS and the calculated TDS from the ion concentrations.
Best used in a dataframe.
}
\examples{
data <- data \%>\% pmap_drf(balance_ions)

}
