% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acid_Base.R
\name{calculate_hardness}
\alias{calculate_hardness}
\title{Hardness Calculation}
\usage{
calculate_hardness(Ca.Conc, Mg.Conc)
}
\arguments{
\item{Ca.Conc}{calcium concentration mg/L}

\item{Mg.Conc}{magnesium concentration mg/L}
}
\description{
This function takes Calcium and Magnesium concentrations and calculates total hardness
}
