% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acid_Base.R
\name{calculate_tds}
\alias{calculate_tds}
\title{TDS Calculation}
\usage{
calculate_tds(Alk, Na.Conc, Ca.Conc, Mg.Conc, K.Conc, Cl.Conc, SO4.Conc)
}
\arguments{
\item{Alk}{Alkalinity in mg as CacO3/L}

\item{Na.Conc}{sodium concentration in mg/L}

\item{Ca.Conc}{calcium concentration mg/L}

\item{Mg.Conc}{magnesium concentration mg/L}

\item{K.Conc}{potassium concentration mg/L}

\item{Cl.Conc}{chloride concentration mg/L}

\item{SO4.Conc}{sulfate concentration mg/L}
}
\description{
This function takes alkalinity and ions and calculates TDS
}
