% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acid_Base.R
\name{chem_calc_ph}
\alias{chem_calc_ph}
\title{Calculate pH, Alk, and optional corrosivity parameters given raw water quality and optional chemical addition}
\usage{
chem_calc_ph(
  Temp0,
  Alk0,
  pH0,
  Cond = 500,
  TDS = NA,
  IS = NA,
  Temp.Final = Temp0,
  HCl.Conc = 0,
  H2SO4.Conc = 0,
  H3PO4.Conc = 0,
  CO2.Gas = 0,
  NaOH.Conc = 0,
  Na2CO3.Conc = 0,
  NaHCO3.Conc = 0,
  Lime.Conc = 0,
  MgOH2.Conc = 0,
  Cl2.Gas = 0,
  Cl2.Conc = 0,
  CaOCl2.Conc = 0,
  Alum.Conc = 0,
  FerricCl.Conc = 0,
  FerricSO4.Conc = 0,
  CaCO3.Conc = 0,
  Na.Conc = NA,
  Ca.Conc = NA,
  Mg.Conc = NA,
  K.Conc = 0,
  Cl.Conc = NA,
  SO4.Conc = NA,
  ionunits = "mg",
  corrosivity = FALSE,
  ...
)
}
\arguments{
\item{Temp0}{temperature in degrees C}

\item{Alk0}{alkalinity in mg CaCO3/L}

\item{pH0}{pH in SU}

\item{Cond}{conductivity (optional, use when ionic strength and ion concentrations are unknown)}

\item{TDS}{total dissolved solids (optional, use when IS and ion concentrations are unknown)}

\item{IS}{ionic strength (optional, enter if known)}

\item{Temp.Final}{final temperature (optional, enter if different from initial)}

\item{HCl.Conc}{Hydrochloric Acid | HCl    | HCl -> H+ + Cl- (all chemical doses default to 0, units are mg/L)}

\item{H2SO4.Conc}{Sulfuric Acid | H2SO4    | H2SO4  -> 2H+ + SO42-}

\item{H3PO4.Conc}{Phosphoric Acid | H3PO4    | H3PO4  -> 3H+ + PO43-}

\item{CO2.Gas}{Carbon Dioxide | CO2 (gas) | CO2 + H2O -> H2CO3*}

\item{NaOH.Conc}{sodium hydroxide/Caustic Soda  | NaOH    | NaOH -> Na+ + OH-}

\item{Na2CO3.Conc}{Soda ash  | Na2CO3    | Na2CO3  -> 2Na+ + CO32-}

\item{NaHCO3.Conc}{Sodium Bicarbonate | NaHCO3    | NaHCO3  -> Na+ + H+ + CO32-}

\item{Lime.Conc}{Lime | Ca(OH)2    | Ca(OH)2 -> Ca2+ + 2OH-}

\item{MgOH2.Conc}{Magnesium Hydroxide | Mg(OH)2    | Mg(OH)2 -> Mg2+ + 2OH-}

\item{Cl2.Gas}{Chlorine Gas  | Cl2 (gas)    | Cl2 + H2O -> HOCl + H+ + Cl-}

\item{Cl2.Conc}{Sodium Hypochlorite | Cl2 | NaOCl    NaOCl -> Na+ + OCl-}

\item{CaOCl2.Conc}{Calcium Hypochlorite (HTH) | Cl2  | Ca(OCl)2    Ca(OCl)2 -> Ca2+ + 2OCl-}

\item{Alum.Conc}{Alum  | Al2(SO4)3\emph{14H2O    | Al2(SO4)3}14H2O + 6HCO3- -> 2Al(OH)3(am) +3SO42- + 14H2O + 6CO2}

\item{FerricCl.Conc}{Ferric Chloride | FeCl3    | FeCl3 +3HCO3- -> Fe(OH)3(am) + 3Cl- + 3CO2}

\item{FerricSO4.Conc}{Ferric Sulfate | Fe2(SO4)3    | Fe2(SO4)3 + 6HCO3- -> 2Fe(OH)3(am) +3SO42- + 6CO2}

\item{CaCO3.Conc}{Calcium Carbonate | CaCO3 | CaCO3 -> Ca2+ + CO32-}

\item{Na.Conc}{sodium concentration (optional, enter if known)}

\item{Ca.Conc}{calcium concentration (optional, enter if known)}

\item{Mg.Conc}{magnesium concentration (optional, enter if known)}

\item{K.Conc}{potassium concentration (optional, defaults to 0)}

\item{Cl.Conc}{chloride concentration (optional, enter if known)}

\item{SO4.Conc}{sulfate concentration (optional, enter if known)}

\item{ionunits}{if Na, Ca, Mg, K, Cl, and SO4 are not entered in mg/L, provide the units here: "mol", "eq", or "mgCaCO3" (per L)}

\item{corrosivity}{if TRUE, corrosivity parameters (LSI, Delta G, DIC, CSMR, Larson SKold index, AI, and RI) will be included in output.}
}
\description{
Given raw pH, Alk, Temp and chem doses, find pH and alk (corrosivity params optional). Works well in a dataframe,
but this will require/create a lot of columns.
}
\examples{
chem_calc_ph(20, 120, 8.1, Alum.Conc = 40)
chem_calc_ph(20, 120, 8.1, Cond = 350, H2SO4.Conc = 8, FerricSO4.Conc = 20, corrosivity = TRUE)

}
