% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acid_Base.R
\name{convert}
\alias{convert}
\title{Unit conversions}
\usage{
convert(Conc, Compound, end = "mol", start = "mg")
}
\arguments{
\item{Conc}{concentration of the compound in the current units}

\item{Compound}{chemical formula for the compound to be converted in quotes. Currently accepts:
"Na", "K", "Cl", "HCO3", "Ca", "Mg", "SO4", "CO3"}

\item{end}{desired units/L of conversion in quotes. Currently accepts: "mg", "mol", "eq", "mgCaCO3"}

\item{start}{starting units/L of conversion in quotes. Currently accepts: "mg", "mol", "eq", "mgCaCO3"}
}
\description{
This function does unit conversions of common ions.
}
\examples{
convert(50, "Ca", end = "mgCaCO3", start = "mg")

}
