% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Useful_Fcns.R
\name{corona_boxplot}
\alias{corona_boxplot}
\title{Custom box plot function that uses the 5th, 25th, 50th, 75th, and 95th quantiles}
\usage{
corona_boxplot(
  outlier = TRUE,
  cdist = 0,
  hbar = TRUE,
  count = TRUE,
  diamond = "mean",
  fill = NA,
  alpha = 0.8,
  pointsize = 1,
  fontsize = 9,
  width = 0.9,
  overlap = FALSE
)
}
\arguments{
\item{outlier}{Default is that outlier points are included.
The n count will be placed under the minimum outlier point if outliers are included.
If outliers are not included, the n count will be placed under the 25th percentile whisker.}

\item{cdist}{The distance n count will be from either the minimum outlier point or the 25th percentile whisker.}

\item{hbar}{Horizontal bars on the whiskers.}

\item{count}{Display count below the box and whisker. Default displays count.}

\item{diamond}{Diamond can be equal to "mean" or "90th" to place the point at the mean or 90th percentile.
If set to FALSE, no diamond point will plot.}

\item{fill}{Can be called to manually set one fill color for the box and diamond. Use corona_hex_list to specify color.}

\item{alpha}{Can be used to manually set one alpha for the boxes (defaults to 0.8). If an alpha scale is required, set to NA}

\item{pointsize}{Scales the points (diamond and outlier)}

\item{fontsize}{Can be used to change the size of the count}

\item{width}{Can be used to change the width of the boxes}

\item{overlap}{When set to TRUE, the position_identity function is called on all elements to place all boxes on top of each other.}
}
\description{
To use this function, you need to call a ggplot() first and give it the data and aesthetic arguments
i.e. ggplot(Data, aes(x = , y = , fill = ))
}
\examples{
gacbox <- ggplot(WQ.TCP, aes(x=Site.Name, y= Calc.Result)) +
 corona_boxplot(cdist = 70, fill = corona_hex_list$navy, alpha = .8)

}
