% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Useful_Fcns.R
\name{corona_boxplot_legend}
\alias{corona_boxplot_legend}
\alias{chad}
\title{Legend for Corona custom box plot}
\usage{
corona_boxplot_legend(
  outlier = TRUE,
  hbar = TRUE,
  count = TRUE,
  diamond = "mean",
  alpha = 0.8,
  fill = corona_hex_list$navy,
  fontsize = 10,
  legendwidth = 1
)

chad(
  outlier = TRUE,
  hbar = TRUE,
  count = TRUE,
  diamond = "mean",
  alpha = 0.8,
  fill = corona_hex_list$navy,
  fontsize = 10,
  legendwidth = 1
)
}
\arguments{
\item{outlier}{Default is that outlier points are included.
The n count will be placed under the minimum outlier point if outliers are included.
If outliers are not included, the n count will be placed under the 25th percentile whisker.}

\item{hbar}{Horizontal bars on the whiskers.}

\item{diamond}{Diamond can be equal to "mean" or "90th" to place the point at the mean or 90th percentile.
If set to FALSE, no diamond point will plot.}

\item{alpha}{Sets the transparency of the box}

\item{fill}{Can be called to manually set one fill color for the box and diamond. Use corona_hex_list to specify color.
Default is Corona blue, fill = NA will make an unfilled box.}

\item{fontsize}{Can be adjusted to change font size - may have to iterate when you export the plot}

\item{legendwidth}{Can be adjusted to change the width of the box.  Increase the number to increase the box width.}
}
\value{

}
\description{
This function alone will create a ggplot object that can be grouped with a box and whisker plot using draw_plot from cowplot.
}
\examples{
boxlegend <- corona_boxplot_legend(fontsize = 9)
fullbox <- ggdraw(boxandwhisker) + draw_plot(boxlegend, x = .65, y = .6, width = .35, height = .4)

fullbox <- ggdraw(boxandwhisker) + draw_plot(chad(x = .65, y = .6, width = .35, height = .4))

}
