% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Useful_Fcns.R
\name{corona_gdrive}
\alias{corona_gdrive}
\title{Creates GDrive filepath
The function generates the correct file path for mac or PC}
\usage{
corona_gdrive(project_name = "")
}
\arguments{
\item{project_name}{Put the project name that matches the name in the G Drive}
}
\description{
Creates GDrive filepath
The function generates the correct file path for mac or PC
}
\examples{
corona_gdrive("HBWS TCP Study")

}
