% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Useful_Fcns.R
\docType{data}
\name{corona_hex_list}
\alias{corona_hex_list}
\title{Corona Color Hex Codes
If you need to manually set colors use this function}
\format{
An object of class \code{list} of length 10.
}
\usage{
corona_hex_list
}
\description{
Corona Color Hex Codes
If you need to manually set colors use this function
}
\examples{
coronahexlist$navy

}
\keyword{datasets}
