% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Useful_Fcns.R
\name{corona_png}
\alias{corona_png}
\title{Generate png}
\usage{
corona_png(plot, filename, scale = 1)
}
\arguments{
\item{plot}{The plot that you are exporting (What it is called in R)}

\item{filename}{The name you want the plot to export as}

\item{scale}{The multiplier you can change the defualt height to. The default height is 4 in,
add the multiplier to change it to what you want}
}
\description{
Generate png
}
\examples{
corona_png("Historical Water Quality Data", wqdataplot, scale = 1.5)

}
