% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Useful_Fcns.R
\name{corona_theme}
\alias{corona_theme}
\title{Custom Corona plot formatting (includes the color palette)}
\usage{
corona_theme(color = TRUE)
}
\arguments{
\item{color}{This determines whether the scale_color_corona and scale_fill_corona functions are called as
part of the theme.  If set to FALSE, they will not be included.  Use if you need a continuous color or fill scale.}
}
\description{
This formatting sets the text size, however feel free to override when needed using (here are options, pick one at a time to change the size of):
axis.text, axis.title, legend.text, legend.title = element_text(size = )
}
\examples{
ggplot(data) + geom_point(aes(x = Date, y = Result, color = Site)) + corona_theme()

}
