% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots_Useful_Fcns.R
\name{scale_fill_corona}
\alias{scale_fill_corona}
\title{Corona color palette (fill)}
\usage{
scale_fill_corona(palette = "sunset", ...)
}
\arguments{
\item{palette}{"sunset" is currently the only option with Corona colors (and is the default).
Using Town of Gilbert colors, you can call:
"g_primary", "g_cool", "g_warm", "g_dark", "g_all", "g_babyshark", "g_mamashark", "g_narwhal".
To add more custom color palettes, the source code for this package must be modified.}
}
\description{
Corona color palette (fill)
}
\examples{
ggplot(data) + geom_bar(aes(x = Site, y = Result, color = Parameter)) + scale_fill_corona()
}
