% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Acid_Base.R
\name{solve_chem_dose}
\alias{solve_chem_dose}
\title{Calculate chemical dose based on target pH}
\usage{
solve_chem_dose(
  Temp0,
  Alk0,
  pH0,
  pH.Final,
  Cond = 500,
  TDS = NA,
  IS = NA,
  Na.Conc = NA,
  Ca.Conc = NA,
  Mg.Conc = NA,
  K.Conc = NA,
  Cl.Conc = NA,
  SO4.Conc = NA,
  chemical,
  ionunits = "mg",
  ...
)
}
\arguments{
\item{Temp0}{temperature in degrees C}

\item{Alk0}{alkalinity in mg CaCO3/L}

\item{pH0}{starting pH in SU}

\item{pH.Final}{desired final pH in SU}

\item{Cond}{conductivity (optional, use when ionic strength and ion concentrations are unknown)}

\item{TDS}{total dissolved solids (optional, use when ionic strength and ion concentrations are unknown)}

\item{IS}{ionic strength (optional, enter if known)}

\item{Na.Conc}{sodium concentration (optional, enter if known)}

\item{Ca.Conc}{calcium concentration (optional, enter if known)}

\item{Mg.Conc}{magnesium concentration (optional, enter if known)}

\item{K.Conc}{potassium concentration (optional, defaults to 0)}

\item{Cl.Conc}{chloride concentration (optional, enter if known)}

\item{SO4.Conc}{sulfate concentration (optional, enter if known)}

\item{chemical}{enter name of chemical to be solved for in quotes, caustic, lime, hydrochloric, CO2, or sulfuric}

\item{ionunits}{if Na, Ca, Mg, K, Cl, and SO4 are not entered in mg/L, provide the units here: "mol", "eq", or "mgCaCO3" (per L)}
}
\description{
Given pH, alk, a target pH, and chemical name, solve for target dose
}
\examples{
solve_chem_dose(20, 120, 6.8, 8.0)

}
